Turn subdomain name into link in DNS page.
This is a quick solution for #13. This has 1 side effect and 1 potentially undesirable behavior. First of all, it adds an underline to the subdomain link. While this is easily fixed, I prefer to make the least amount of changes possible. Additionally, since it is a link, an underline may be appropriate. Second of all, it links with http and not https. This is due to us not knowing if the website is https compatible and playing it safe. Without keeping track if all the subdomains are https compatible, this is the best solution that I know of.